requirements.txt
python-dotenv==1.2.1
pyyaml==6.0.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
google-generativeai==0.3.0
requests==2.31.0
requests-oauthlib==1.3.0
click==8.1.7
